test: add comprehensive Phase 1 test suite (180 tests)
- Set up Vitest workspace with 6 projects (shared, adapter-*, api-unit, api-integration)
- Add test helpers: mock-db (Proxy-based Drizzle mock), fixtures, test-app, test-db, global-setup
- Add shared unit tests: errors, validators (package, org, group, resource) — 52 tests
- Add adapter unit tests: in-process queue, noop AI, local storage — 19 tests
- Add service unit tests: package, org, group, resource, tag services — 43 tests
- Add middleware unit test: error-handler RFC 7807 conversion — 5 tests
- Add integration tests: packages, orgs, groups, resources, tags, CKAN compat, health — 66 tests
- Extract isUuid() to @kukan/shared, fix getByNameOrId consistency across services
- Fix OrganizationService UUID detection bug (was causing DB type error)
- Fix TagService COUNT returning string instead of number (add ::int cast)
- Remove custom Postgres Dockerfile, use postgres:16-alpine image directly

+++ apps/api/src/services/tag-service.ts
@@ -23,7 +23,7 @@ export class TagService {
         id: tag.id,
         name: tag.name,
         vocabularyId: tag.vocabularyId,
-        packageCount: sql<number>`COUNT(DISTINCT ${packageTag.packageId})`.as('package_count'),
+        packageCount: sql<number>`COUNT(DISTINCT ${packageTag.packageId})::int`.as('package_count'),
       })
       .from(tag)
       .leftJoin(packageTag, eq(tag.id, packageTag.tagId))
@@ -46,7 +46,7 @@ export class TagService {
         id: tag.id,
         name: tag.name,
         vocabularyId: tag.vocabularyId,
-        packageCount: sql<number>`COUNT(DISTINCT ${packageTag.packageId})`.as('package_count'),
+        packageCount: sql<number>`COUNT(DISTINCT ${packageTag.packageId})::int`.as('package_count'),
       })
       .from(tag)
       .leftJoin(packageTag, eq(tag.id, packageTag.tagId))
